+++ lib/Redesign/Testin_components/building/building.dart
@@ -8,79 +8,76 @@ class Building extends StatelessWidget {
     return const Scaffold(
       body: Row(
         children: [
-          SizedBox(width: 600, child: Listing()),
-          SizedBox(width: 600, child: Withclass()),
+          SizedBox(child: CreateRecipeListCard2()),
         ],
       ),
     );
   }
 }
 
-class Listing extends StatefulWidget {
-  const Listing({super.key});
-
-  @override
-  State<Listing> createState() => _ListingState();
-}
-
-class _ListingState extends State<Listing> {
-  // Sample data
-  final List<Map<String, String>> _potatis = [
-    {'label': 'Item 1', 'amount': '100', 'unit': 'kg'},
-    {'label': 'Item 2', 'amount': '200', 'unit': 'g'},
-    // Add more items as needed
-  ];
+class CreateRecipeListCard2 extends StatelessWidget {
+  const CreateRecipeListCard2({
+    super.key,
+  });
 
   @override
   Widget build(BuildContext context) {
-    return ReorderableListView(
-      children: List.generate(_potatis.length, (index) {
-        Map<String, String> item = _potatis[index];
-        print('potatis');
-        return EditingTile(
-          key: ValueKey('item_${item['label']}'),
-          label: item['label']!,
-          amount: item['amount']!,
-          unit: item['unit']!,
-        );
-      }),
-      onReorder: (int oldIndex, int newIndex) {
-        setState(() {
-          if (newIndex > oldIndex) {
-            newIndex -= 1;
-          }
-          final item = _potatis.removeAt(oldIndex);
-          _potatis.insert(newIndex, item);
-        });
-      },
+    return SizedBox(
+      width: 500,
+      height: 450,
+      child: Card(
+        color: const Color(0xffd9d9d9),
+        child: ClipRRect(
+          borderRadius: BorderRadius.circular(10),
+          child: const Column(
+            children: [
+              CreateIngMainTile(),
+              Expanded(
+                child: ReorderIngridientlist(),
+              ),
+            ],
+          ),
+        ),
+      ),
     );
   }
 }
 
-class Withclass extends StatefulWidget {
-  const Withclass({super.key});
+class ReorderIngridientlist extends StatefulWidget {
+  const ReorderIngridientlist({super.key});
 
   @override
-  State<Withclass> createState() => _WithclassState();
+  State<ReorderIngridientlist> createState() => _ReorderIngridientlistState();
 }
 
-class _WithclassState extends State<Withclass> {
+class _ReorderIngridientlistState extends State<ReorderIngridientlist> {
   final List<Item> _items = [
-    Item(label: 'Item 1', amount: 100, unit: 'kg'),
-    Item(label: 'Item 2', amount: 200, unit: 'g'),
+    Item(label: 'maricha sås', amount: '300', unit: 'ml'),
+    Item(label: 'Jordgubbssylt', amount: '4', unit: 'dl'),
+    Item(label: 'Kebabkött', amount: '400', unit: 'g'),
+    Item(label: 'vitlöksås', amount: '1', unit: 'l'),
+    Item(label: 'vitlök', amount: '2', unit: 'st'),
+    Item(label: 'gullök', amount: '2000', unit: 'tsk'),
+    Item(label: 'honung', amount: '1000', unit: 'ml'),
+    Item(label: 'socker', amount: '2000', unit: 'tsk'),
+    Item(label: 'mjölk', amount: '1000', unit: 'ml'),
+    Item(label: 'Bröd', amount: '2000', unit: 'tsk'),
+    Item(label: 'Fetaost', amount: '1000', unit: 'ml'),
+    Item(label: 'Ketchup', amount: '2000', unit: 'tsk'),
     // ... more items
   ];
-
   @override
   Widget build(BuildContext context) {
     return ReorderableListView(
+      buildDefaultDragHandles: false,
       children: List.generate(_items.length, (index) {
         Item item = _items[index];
-        return EditingTile(
+        return Tilebbs(
           key: ValueKey(item.label),
           label: item.label,
-          amount: item.amount.toString(), // Convert int to String
+          amount: item.amount,
           unit: item.unit,
+          index: index,
         );
       }),
       onReorder: (int oldIndex, int newIndex) {
@@ -89,6 +86,7 @@ class _WithclassState extends State<Withclass> {
             newIndex -= 1;
           }
           final Item item = _items.removeAt(oldIndex);
+
           _items.insert(newIndex, item);
         });
       },
@@ -98,30 +96,83 @@ class _WithclassState extends State<Withclass> {
 
 class Item {
   final String label;
-  final int amount;
+  final String amount;
   final String unit;
 
   Item({required this.label, required this.amount, required this.unit});
 }
 
-final List<Item> _items = [
-  Item(label: 'Item 1', amount: 100, unit: 'kg'),
-  Item(label: 'Item 2', amount: 200, unit: 'g'),
-  // ...
-];
+class CreateIngMainTile extends StatelessWidget {
+  const CreateIngMainTile({
+    super.key,
+  });
 
-class EditingTile extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    const textStyle = TextStyle(
+      fontSize: 24,
+      fontWeight: FontWeight.w700,
+      color: Color(0xFF000000), // Set your desired color here
+    );
+
+    return Container(
+      height: 50,
+      width: 500,
+      color: const Color(0xFFD9D9D9),
+      child: const Stack(
+        children: [
+          Positioned(
+            left: 20,
+            top: 10,
+            child: carticon(),
+          ),
+          Positioned(
+            top: 10,
+            left: 0,
+            right: 0,
+            child: Align(
+              alignment: Alignment.center,
+              child: Text("Ingridients", style: textStyle),
+            ),
+          ),
+          Positioned(
+            top: 10,
+            right: 20,
+            child: carticon(),
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+class carticon extends StatelessWidget {
+  const carticon({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Image.asset('images/new/icons/cart.png');
+  }
+}
+
+class Tilebbs extends StatelessWidget {
   final String label; // First parameter
   final String amount;
   final String unit;
+  final int index;
+
   // Third parameter
 
   // Update the constructor to accept three parameters
-  const EditingTile({
+  const Tilebbs({
     super.key,
+    // Initialize key
     required this.label,
     required this.amount,
     required this.unit,
+    required this.index, // Initialize index
   });
 
   @override
@@ -134,46 +185,90 @@ class EditingTile extends StatelessWidget {
 
     return Container(
       height: 50,
-      width: 410,
+      width: 500,
       color: const Color(0xff161414),
       child: SizedBox(
-        child: Row(
+        child: Stack(
           children: [
-            const SizedBox(
-              width: 16,
-            ),
-            IconButton(
-              onPressed: () {
-                print('menue clicked');
-              },
-              icon: const Icon(
-                Icons.menu,
-                color: Color.fromARGB(255, 233, 228, 228),
+            Positioned(
+              top: 5,
+              left: 5,
+              child: ReorderableDragStartListener(
+                index: index,
+                child: IconButton(
+                  onPressed: () {
+                    print('menue clicked');
+                  },
+                  icon: const Icon(
+                    Icons.menu,
+                    color: Color.fromARGB(255, 233, 228, 228),
+                  ),
+                ),
               ),
             ),
-            const SizedBox(width: 16),
-            Text(label, style: textStyle), // Use the label parameter
-            const Text(':', style: textStyle),
-            const SizedBox(width: 20),
-            Text(amount, style: textStyle), // Use the quantity parameter
-            const SizedBox(width: 3),
-            Text(unit, style: textStyle), // Use the unit parameter
-            const SizedBox(
-              width: 130,
-            ),
-            IconButton(
-              constraints: const BoxConstraints(maxHeight: 100, maxWidth: 100),
-              icon: const Icon(Icons.edit, size: 24),
-              color: const Color.fromARGB(255, 233, 228, 228),
-              // Plus icon
-              onPressed: () {
-                print('Edit clicked');
-                // Add your action for this button
-              },
+            Positioned(
+                top: 5,
+                left: 45,
+                child: Text('$label:',
+                    style: textStyle)), // Use the label parameter
+            Positioned(
+                top: 5,
+                left: 300,
+                child: Text('$amount $unit',
+                    style: textStyle)), // Use the quantity parameter
+
+            // Use the unit parameter
+
+            const Positioned(
+              top: 5,
+              right: 5,
+              child: Row(
+                children: [Editcontainer(), Deletecontainer()],
+              ),
             ),
           ],
         ),
       ),
+    );
+  }
+}
+
+class Deletecontainer extends StatelessWidget {
+  const Deletecontainer({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return IconButton(
+      constraints: const BoxConstraints(maxHeight: 100, maxWidth: 100),
+      icon: const Icon(Icons.delete, size: 24),
+      color: const Color.fromARGB(255, 233, 228, 228),
+      // Plus icon
+      onPressed: () {
+        print('delete button');
+        // Add your action for this button
+      },
+    );
+  }
+}
+
+class Editcontainer extends StatelessWidget {
+  const Editcontainer({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return IconButton(
+      constraints: const BoxConstraints(maxHeight: 100, maxWidth: 100),
+      icon: const Icon(Icons.edit, size: 24),
+      color: const Color.fromARGB(255, 233, 228, 228),
+      // Plus icon
+      onPressed: () {
+        print('Edit clicked');
+        // Add your action for this button
+      },
     );
   }
 }
