--- lib/Redesign/Testin_components/building/reoarder.txt
@@ -1,17 +1,0 @@
-
-class Listing extends StatefulWidget {
-  const Listing({super.key});
-
-  @override
-  State<Listing> createState() => _ListingState();
-}
-
-class _ListingState extends State<Listing> {
-  @override
-  Widget build(BuildContext context) {
-    return ReorderableListView(children: [
-
-    ],
-    onReorderEnd: ,);
-  }
-}
