+++ lib/Redesign/Testin_components/central_redirect/Building _Central/building_redirected.dart
@@ -20,10 +20,6 @@ class BuildingRedirect extends StatelessWidget {
                 padding: EdgeInsets.symmetric(horizontal: 16.0),
                 child: BuildingActiveButton(),
               ),
-              Padding(
-                padding: EdgeInsets.symmetric(horizontal: 16.0),
-                child: B2Button(),
-              ),
             ],
           ),
         ),
